Cleaned up test code
Split off part of the test_the.cpp code into test_playback.cpp, as it was actually testing the playback and serialization functionality.  Added another functional test case for the THE game loop and organized the test cases using the header file to contain test fixtures data and a template function to load that data into a test case struct and execute the test.